Раскидал по папкам. Удалил мусорный код. Изменил отображение погоды.
