+++ app/src/main/java/ru/oliverhd/weather/CityAdapter.java
@@ -1,8 +1,11 @@
 package ru.oliverhd.weather;
 
+/*
+* Адаптер для RecyclerView городов.
+* */
+
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
