--- app/src/main/java/ru/oliverhd/weather/WeatherDetailFragment.java
@@ -1,118 +1,0 @@
-package ru.oliverhd.weather;
-
-import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import android.os.Handler;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import com.google.gson.Gson;
-
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.stream.Collectors;
-
-import javax.net.ssl.HttpsURLConnection;
-
-import ru.oliverhd.weather.interfaces.Constants;
-import ru.oliverhd.weather.model.WeatherRequest;
-
-public class WeatherDetailFragment extends Fragment implements Constants {
-
-    private static final String TAG = "Weather";
-
-    public static WeatherDetailFragment create(Parcel parcel) {
-        WeatherDetailFragment fragment = new WeatherDetailFragment();
-        Bundle args = new Bundle();
-        args.putSerializable(PARCEL, parcel);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    public Parcel getParcel() {
-        Parcel parcel = (Parcel) getArguments().getSerializable(PARCEL);
-        return parcel;
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_weather_detail, container, false);
-    }
-
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-
-        final TextView textView = view.findViewById(R.id.weather_info);
-        TextView city = view.findViewById(R.id.city);
-        Parcel parcel = getParcel();
-        city.setText(parcel.getCityName());
-        try {
-            final String url = String.format("https://api.openweathermap.org/data/2.5/weather?q=%s&appid=%s", city.getText(), BuildConfig.WEATHER_API_KEY);
-            final URL uri = new URL(url);
-            final Handler handler = new Handler();
-            new Thread(new Runnable() {
-                public void run() {
-                    HttpsURLConnection urlConnection = null;
-                    try {
-                        urlConnection = (HttpsURLConnection) uri.openConnection();
-                        urlConnection.setRequestMethod("GET");
-                        urlConnection.setReadTimeout(10000);
-
-                        BufferedReader in = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
-                        String result = getLines(in);
-
-                        Gson gson = new Gson();
-                        final WeatherRequest weatherRequest = gson.fromJson(result, WeatherRequest.class);
-                        handler.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                textView.setText(Float.toString((weatherRequest.getMain().getTemp() - 273)));
-                            }
-                        });
-
-
-                    } catch (Exception e) {
-                        Log.e(TAG, "Fail connection", e);
-                        e.printStackTrace();
-                    } finally {
-                        if (urlConnection != null) {
-                            urlConnection.disconnect();
-                        }
-                    }
-                }
-            }).start();
-        } catch (MalformedURLException e) {
-            Log.e(TAG, "Fail URI", e);
-            e.printStackTrace();
-        }
-
-        String[] data = getResources().getStringArray(R.array.time);
-
-        RecyclerView recyclerView = view.findViewById(R.id.days_recycler);
-//        GridLayoutManager gridLayout = new GridLayoutManager(getActivity(), GridLayoutManager.DEFAULT_SPAN_COUNT, LinearLayoutManager.HORIZONTAL, false);
-        LinearLayoutManager layoutManager = new LinearLayoutManager(getActivity(), LinearLayoutManager.HORIZONTAL, false);
-        recyclerView.setLayoutManager(layoutManager);
-
-        recyclerView.setHasFixedSize(true);
-
-        DetailAdapter adapter = new DetailAdapter(data);
-        recyclerView.setAdapter(adapter);
-    }
-
-    private String getLines(BufferedReader in) {
-        return in.lines().collect(Collectors.joining("\n"));
-    }
-}
