+++ app/src/main/java/ru/oliverhd/weather/interfaces/OpenWeather.java
@@ -3,7 +3,7 @@ package ru.oliverhd.weather.interfaces;
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
-import ru.oliverhd.weather.model.WeatherRequest;
+import ru.oliverhd.weather.json_model.weather.WeatherRequest;
 
 public interface OpenWeather {
