+++ app/src/main/res/layout-land/activity_main.xml
@@ -18,7 +18,7 @@
     </FrameLayout>
 
     <FrameLayout
-        android:id="@+id/weather_info"
+        android:id="@+id/temperature_text_view"
         android:layout_width="0px"
         android:layout_height="match_parent"
         android:layout_weight="1" />
