+++ app/src/main/res/layout/fragment_cities.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".CitiesFragment">
+    tools:context=".fragments.CitiesFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
