+++ app/src/main/res/layout/fragment_weather_detail.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WeatherDetailFragment">
+    tools:context=".fragments.WeatherDetailFragment">
 
     <TextView
-        android:id="@+id/weather_info"
+        android:id="@+id/temperature_text_view"
         style="@style/textViewStyle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -31,7 +31,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="100dp"
-        app:layout_constraintTop_toBottomOf="@+id/weather_info">
+        app:layout_constraintTop_toBottomOf="@+id/temperature_text_view">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/days_recycler"
