+++ app/src/main/res/menu/main_menu.xml
@@ -2,10 +2,15 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <item
+        android:id="@+id/action_history"
+        android:orderInCategory="100"
+        android:title="@string/history"
+        app:showAsAction="never" />
 
     <item
         android:id="@+id/action_settings"
-        android:orderInCategory="100"
+        android:orderInCategory="50"
         android:title="@string/settings"
         app:showAsAction="never" />
